Update docker compose process PROJECT_MODE='dev' -Add new env. vars APP_USER_ID, APP_GROUP_ID, and APP_GROUP_NAME -Modify docker-compose-dev+migrations.sh to: a. Call 'docker compose' instead of 'docker-compose' for cross-platform compatibility b. Verify that the above-mentioned env. vars along with LOGS_DIR_NAME are properly set in the_wall_api_dev.env c. Use them to create the app's logs folder
Linux-only:
d. On the host, change ownership for the mounted logs dir to the newly
created app group
e. Modify the permissions management inside the docker containers:
Create an app group with the same GID and name as the one described
above, add the non-root app user to it, and set it as owner to the
mounted logs directory